Homework22
fixed previous issues
