--- src/test/java/Tests/BaseTest1.java
@@ -1,38 +1,0 @@
-package Tests;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.BeforeSuite;
-import org.testng.annotations.Parameters;
-
-import java.time.Duration;
-
-public class BaseTest1
-{
-    public static WebDriver driver = null;
-    public static String url = null;
-    public static WebDriverWait wait = null;
-    public static Actions actions = null;
-    public static ChromeOptions options = null;
-    @BeforeSuite
-    static void setupClass() { WebDriverManager.chromedriver().setup();}
-
-    @BeforeMethod
-    @Parameters({"baseURL"})
-    public void launchBrowser(String baseURL)
-    {
-        options = new ChromeOptions();
-        options.addArguments("--remote-allow-origins=*");
-        options.addArguments("--start-maximized");
-        options.addArguments("--disable-notifications");
-        driver=new ChromeDriver(options);
-        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-        driver.get(baseURL);
-    }
-}
