+++ src/test/resources/TestNG.xml
@@ -5,9 +5,9 @@
     <parameter name="baseURL" value="https://qa.koel.app/"/>
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Tests.BaseTest1"/>
+            <class name="BaseTest"/>
             <class name="LoginTests"/>
-            <class name="Pages.HomeTests"/>
+            <class name="HomeTests"/>
         </classes>
     </test>
 </suite>
